renaming the queue file to avoid import conflict with snowflake connector package
