+++ q.py
